Add the --no-iri flag to wget-at arguments.
When downloading repository files, bintray redirects us to one of
serveral CDNs. One of which is cloudfront.
If a file has a ~ in its name, like https://dl.bintray.com/lamyj/generic/apt/pool/main/o/odil/python3-odil_0.12.1-1~stretch_amd64.deb
then the url in the Location header from bintray will be to cloudfront
with %7E in it (url escaped version of ~).
wget-at will then modify the url in the location header and unencode the
%7E before making a request to cloudfront.
Cloudfront does not consider the file with the ~ in it to be equal to
the escaped %7E version.
As such we need to disable this unescaping behaviour.
The --no-iri flag seems, from my testing, to disable this behaviour and
allows the file to download correctly.